Add files via upload
The string should be entered using input() and assigned to a variable. 
Use print() twice at the end of your program.  The first print() will print the string that the user entered and the second print() will display the number of characters in the string.  For example, if the user entered the string "hello world", the number of characters would be 11.