Add files via upload
Create a variable called the_string and assign it the string "North India".
Call .rjust() on the_string with 15 as its argument and print() the result.
Call .ljust() on the_string with the arguments 15 and "*" then print() the result.
Create a variable called center_plus and assign it the result of .center() being called on the_string with 14 and "+" as arguments.
Use print() to display the string assigned to center_plus.
Call .lstrip() on the_string to remove "North" then print() the result.
Call .rstrip() on center_plus with "+" as its argument and print() the result.
Call .strip() on center_plus with "+" as its argument and print() the result.
Call .replace() on the_string and replace "North" with "South".  print() the result.