Add files via upload
The program utelizes a for loop to go through each character in a string and determine if it is already present in a dictionary. If it is, the value of that character is incremented by 1. If not, the character is added to the dictionary with a value of 1. The end result is a list of all unique characters found in the given string. The dictionary is created with the string itself as the key and a value of 1. As the loop goes through each character in the string, it checks if it already exists in the dictionary. If not, a new entry is added with the character as the key and a value of 1.