Add files via upload
This code print the value of NCR,When the input function is given two integers, it converts them into variables n and r respectively. The value of r must be positive and less than or equal to n. Using these values, the program calculates a specific formula involving factorials and returns the result.