Add files via upload
A python program for a function takes two values and returns the result of dividing them. An assert statement checks if the input value for the first value is greater than 0. If not, the program prints an error message and exits with a code of 0. A variable called "c" holds the result of dividing the two values, with an exception clause in case of errors. If no exceptions occur, "c" is returned as the output value. The code returns "a/b".