Add files via upload
The Body Mass Index (BMI) is a measure of body fat based on height and weight. It is commonly used to classify underweight, overweight, and obesity in adults. The formula for calculating BMI is:

BMI = weight (kg) / height (m)^2

This Python code calculates BMI of a body using the above formula.