Add files via upload
This code creates four lists and two sets. It converts list2 into set1 and updates it with the elements of set2, which is created by converting list1 into a set. Then, it updates set1 with the elements of list3 and list4 using the update() method. The final output is the updated set1.