Add files via upload
The code defines a function `addToInventory` that adds items to an inventory dictionary, increments the count of existing items. It also initializes an inventory dictionary and a list of items to be added to the inventory, then calls the `addToInventory` function with these parameters and stores the updated inventory dictionary back in `inv`. Lastly, it introduces a function `displayInventory` that prints out the details of the inventory.