Add files via upload
The code asks the user to input the name of a job applicant, their college major, their current occupation, and the number of years they have been working in that field. Then, it prints out a sentence that includes all of that information using both the `.format()` method and string concatenation.