Add files via upload
Create a list of lists and print the first list. Print a specific element from the list. Create another list and print an element from it using negative index. Concatenate elements from two lists to form a sentence. Create a list of floating-point numbers and print specific slices from it.