Add files via upload
In this python code, a dictionary containing five pairs of keys and their corresponding values is created. The value associated with the key "c" is printed, followed by checking if the key "a" exists in the dictionary. Lastly, it checks if the key "b" does not exist in the dictionary.