Add files via upload
This Python code , dictionaries are exemplified by generating, modifying, and iterating through them. The `fromkeys()` method is utilized to assign keys, while the `pop()` and `popitem()` methods are employed to eliminate key-value pairs. Additionally, the `print()` function is used to display information on the console. The script presents multiple techniques for managing dictionaries in Python.